requirements.txt
pandas
numpy
scikit-learn
requests
joblib
streamlit
matplotlib
python-dotenv
statsmodels